requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Authentication
pyjwt==2.8.0
cryptography==41.0.7

# Stripe
stripe==7.7.0

# HTTP Client
httpx==0.25.2

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6

# Logging & Monitoring
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Cloud
google-cloud-secret-manager==2.18.0
google-cloud-logging==3.8.0

